# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/courses": {
      "post": {
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "get": {
        "operationId": "CoursesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      }
    },
    "/api/courses/{id}": {
      "get": {
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "patch": {
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseBodyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        },
        "tags": [
          "Courses"
        ]
      },
      "delete": {
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "Courses"
        ]
      }
    },
...
    "/api/academic-periods": {
      "get": {
        "operationId": "AcademicPeriodsController_findAll",
        "parameters": [
          {
            "name": "year",
            "required": false,
            "in": "query",
            "description": "Filtrar etapas por año lectivo",
            "schema": {
              "minimum": 2000,
              "maximum": 2100,
              "example": 2026,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filtrar por nombre de etapa evaluativa",
            "schema": {
              "example": "1ª Etapa",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of academic periods"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List academic periods / stages",
        "tags": [
          "Academic Periods"
        ]
      },
      "post": {
        "operationId": "AcademicPeriodsController_createOrUpdate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAcademicPeriodDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Academic period configured successfully"
          },
          "400": {
            "description": "Validation error (e.g. invalid date order)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Create or update an academic stage definition",
        "tags": [
          "Academic Periods"
        ]
      }
    },
    "/api/academic-periods/years": {
      "get": {
        "operationId": "AcademicPeriodsController_getYears",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of years (numbers)"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "List all distinct academic years with records",
        "tags": [
          "Academic Periods"
        ]
      }
    },
    "/api/academic-periods/{id}": {
      "get": {
        "operationId": "AcademicPeriodsController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Academic period details"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get academic period by ID",
        "tags": [
          "Academic Periods"
        ]
      },
      "put": {
        "operationId": "AcademicPeriodsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAcademicPeriodDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Academic period updated"
          },
          "400": {
            "description": "Invalid date range"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Update stage dates by ID",
        "tags": [
          "Academic Periods"
        ]
      },
      "delete": {
        "operationId": "AcademicPeriodsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Period deleted successfully"
          },
          "400": {
            "description": "Cannot delete period with associated grades"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Period not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Delete academic period by ID",
        "tags": [
          "Academic Periods"
        ]
      }
    },
    "/api/academic-periods/seed/{year}": {
      "post": {
        "operationId": "AcademicPeriodsController_seedDefaults",
        "parameters": [
          {
            "name": "year",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "List of seeded academic periods"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Initialize default Paraguayan academic stages for a year",
        "tags": [
          "Academic Periods"
        ]
      }
    },
    "/api/grades/batch": {
      "post": {
        "operationId": "GradesController_saveBatch",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadGradesBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Grades registered or updated successfully"
          },
          "400": {
            "description": "Validation error in marks or payload"
          },
          "404": {
            "description": "Student or Course not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Batch register or update student evaluations by stage",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/grades/student/{studentId}": {
      "get": {
        "operationId": "GradesController_getHistoryByStudent",
        "parameters": [
          {
            "name": "studentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of student grades"
          },
          "404": {
            "description": "Student not found"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get academic grade history for a student",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/grades/course/{courseId}": {
      "get": {
        "operationId": "GradesController_getGradesByCourse",
        "parameters": [
          {
            "name": "courseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of grades for the course and stage"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get grades for a specific course and stage",
        "tags": [
          "Grades"
        ]
      }
    },
    "/api/grades/stats/averages": {
      "get": {
        "operationId": "GradesController_getAverageGradesByStyle",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Average scores grouped by discipline"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get statistical averages grouped by dance style",
        "tags": [
          "Grades"
        ]
      }
    },
...
  },
  "info": {
    "title": "Yeruti Academia de Danza API",
    "description": "API documentation for the Yeruti Dance Academy management system",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter your JWT token to access secured endpoints",
        "in": "header"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "passwordHash": {
            "type": "string"
          },
          "role": {
            "enum": [
              "Administrator",
              "Docente",
              "Operador",
              "Alumno",
              "Tutor"
            ],
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "passwordHash",
          "role",
          "isActive",
          "firstName",
          "lastName",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CourseSchedule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "course": {
            "$ref": "#/components/schemas/Course"
          },
          "courseId": {
            "type": "string"
          },
          "dayOfWeek": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "classroom": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "course",
          "courseId",
          "dayOfWeek",
          "startTime",
          "endTime",
          "classroom"
        ]
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "capacity": {
            "type": "number"
          },
          "year": {
            "type": "number"
          },
          "classCode": {
            "type": "string",
            "nullable": true
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseSchedule"
            }
          },
          "teacher": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "teacherId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "level",
          "capacity",
          "year",
          "classCode",
          "schedules",
          "teacher",
          "teacherId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CourseScheduleDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "classroom": {
            "type": "string"
          }
        },
        "required": [
          "dayOfWeek",
          "startTime",
          "endTime",
          "classroom"
        ]
      },
      "CreateCourseBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "minimum": 1
          },
          "year": {
            "type": "number"
          },
          "teacherId": {
            "type": "string"
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseScheduleDto"
            }
          }
        },
        "required": [
          "name",
          "level",
          "capacity",
          "schedules"
        ]
      },
      "UpdateCourseBodyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "minimum": 1
          },
          "year": {
            "type": "number"
          },
          "teacherId": {
            "type": "string",
            "nullable": true
          },
          "schedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CourseScheduleDto"
            }
          }
        }
      },
...
      "CreateAcademicPeriodDto": {
        "type": "object",
        "properties": {
          "year": {
            "type": "number",
            "minimum": 2000,
            "maximum": 2100,
            "description": "Año lectivo del ciclo académico",
            "example": 2026
          },
          "name": {
            "type": "string",
            "maxLength": 50,
            "description": "Nombre o denominación de la etapa evaluativa/período",
            "example": "1ª Etapa"
          },
          "startDate": {
            "type": "string",
            "description": "Fecha de inicio de la etapa (YYYY-MM-DD)",
            "example": "2026-02-01"
          },
          "endDate": {
            "type": "string",
            "description": "Fecha de fin de la etapa (YYYY-MM-DD)",
            "example": "2026-06-30"
          }
        },
        "required": [
          "year",
          "name",
          "startDate",
          "endDate"
        ]
      },
      "UpdateAcademicPeriodDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 50,
            "description": "Nombre o denominación de la etapa evaluativa/período",
            "example": "1ª Etapa"
          },
          "startDate": {
            "type": "string",
            "description": "Fecha de inicio de la etapa (YYYY-MM-DD)",
            "example": "2026-02-01"
          },
          "endDate": {
            "type": "string",
            "description": "Fecha de fin de la etapa (YYYY-MM-DD)",
            "example": "2026-06-30"
          }
        }
      },
      "GradeUploadItemDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "format": "uuid",
            "example": "a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11",
            "description": "ID del alumno"
          },
          "courseId": {
            "type": "string",
            "format": "uuid",
            "example": "b0eebc99-9c0b-4ef8-bb6d-6bb9bd380a22",
            "description": "ID del curso (opcional si se define a nivel de lote)"
          },
          "techniqueScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 85,
            "description": "Nota de Técnica (0-100)"
          },
          "expressionScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 90,
            "description": "Nota de Expresión (0-100)"
          },
          "disciplineScore": {
            "type": "number",
            "minimum": 0,
            "maximum": 100,
            "example": 95,
            "description": "Nota de Disciplina (0-100)"
          },
          "stage": {
            "enum": [
              "1ª Etapa",
              "2ª Etapa",
              "Examen Final",
              "Recuperatorio"
            ],
            "type": "string",
            "example": "1ª Etapa",
            "description": "Etapa evaluativa (opcional si se define a nivel de lote)"
          }
        },
        "required": [
          "studentId",
          "techniqueScore",
          "expressionScore",
          "disciplineScore"
        ]
      },
      "UploadGradesBatchDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "string",
            "format": "uuid",
            "example": "b0eebc99-9c0b-4ef8-bb6d-6bb9bd380a22",
            "description": "ID del curso aplicable al lote de calificaciones"
          },
          "stage": {
            "enum": [
              "1ª Etapa",
              "2ª Etapa",
              "Examen Final",
              "Recuperatorio"
            ],
            "type": "string",
            "example": "1ª Etapa",
            "description": "Etapa evaluativa oficial aplicable al lote de calificaciones"
          },
          "grades": {
            "description": "Lista de notas de estudiantes a registrar o actualizar",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GradeUploadItemDto"
            }
          }
        },
        "required": [
          "grades"
        ]
      },
...
    }
  }
}